requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
databases==0.7.0
exceptiongroup==1.1.1
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
pydantic==1.10.7
pypika-tortoise==0.1.6
pytest==7.2.2
python-multipart==0.0.6
pytz==2023.3
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
tomli==2.0.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
